Allow for multiple channels created in a pipeline. Wrangle Catalyst's lazy existance issues. Allow for empty surface MPI ranks.
