+++ pyfr/plugins/soln/insitu/__init__.py
@@ -1,0 +1,7 @@
+from pyfr.plugins.soln.insitu.base import (IntegratorAdapter, bp_key,
+                                           build_cleaner, con_psolns_pgrads,
+                                           face_shape_ops, split_components)
+from pyfr.plugins.soln.insitu.conduit import (ConduitError, ConduitNode,
+                                              ConduitWrappers)
+from pyfr.plugins.soln.insitu.outputs import BoundaryOutput, VolumeOutput
+from pyfr.plugins.soln.insitu.renderer import InSituError, InSituRenderer
